requirements.txt
langchain
langchain-groq
python-dotenv
duckduckgo-search